+++ vista-library/src/main/java/me/oriley/vista/VistaNestedScrollView.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright (C) 2016 Kane O'Riley
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package me.oriley.vista;
+
+import android.content.Context;
+import android.support.annotation.ColorInt;
+import android.support.annotation.NonNull;
+import android.support.v4.widget.NestedScrollView;
+import android.util.AttributeSet;
+import me.oriley.vista.VistaEdgeEffectHelper.Side;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@SuppressWarnings("unused")
+public class VistaNestedScrollView extends NestedScrollView implements VistaEdgeEffectHost {
+
+    private static final String TOP_EDGE = "mEdgeGlowTop";
+    private static final String BOTTOM_EDGE = "mEdgeGlowBottom";
+
+    @NonNull
+    private final VistaEdgeEffectHelper mEdgeEffects;
+
+
+    public VistaNestedScrollView(Context context) {
+        this(context, null);
+    }
+
+    public VistaNestedScrollView(Context context, AttributeSet attrs) {
+        this(context, attrs, 0);
+    }
+
+    public VistaNestedScrollView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+        mEdgeEffects = new VistaEdgeEffectHelper(NestedScrollView.class, this, context, attrs);
+    }
+
+
+    @NonNull
+    @Override
+    public final List<VistaEdgeEffectModel> getEdgeEffectModels() {
+        List<VistaEdgeEffectModel> models = new ArrayList<>();
+        models.add(new VistaEdgeEffectModel(TOP_EDGE, Side.TOP, true));
+        models.add(new VistaEdgeEffectModel(BOTTOM_EDGE, Side.BOTTOM, true));
+        return models;
+    }
+
+    @Override
+    public void setEdgeEffectColors(@ColorInt int color) {
+        mEdgeEffects.setEdgeEffectColors(color);
+    }
+
+    @Override
+    public void setEdgeEffectColor(@NonNull Side side, @ColorInt int color) {
+        mEdgeEffects.setEdgeEffectColor(side, color);
+    }
+
+    @Override
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
+        mEdgeEffects.refreshEdges();
+    }
+}