+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+## Version 0.3.1
+
+_2016-05-08_
+
+ *  Add sample app (core library renamed)
+ *  Bug fixes for RecyclerView/ViewPager edge effects
+
 ## Version 0.3.0
 
 _2016-05-08_